moved decodetoken, started implementing bcrypt
