+++ controllers/users.js
@@ -1,10 +1,11 @@
 require('dotenv').config();
 const db = require('../db/db')
 const jwt = require('jsonwebtoken');
+const bcrypt = require('bcrypt')
 
 const makeToken = (identity) => {
     const accToken = jwt.sign(identity, Buffer.from(process.env.TSK,'base64'),{
-        expiresIn: 700,
+        expiresIn: '1d',
 
     })
     return accToken 
@@ -16,7 +17,6 @@ const login = async(req,res)=> {
         if(user){
             console.log("User found")
             const auth = makeToken({user_id:user.id})
-            res.cookie('token', auth,{httpOnly:true});
             res.status(200).json({token:auth})
         }
         else {
@@ -33,8 +33,9 @@ const login = async(req,res)=> {
 const signUp = async(req,res)=> {
     try{
         console.log(req.body)
-        await db.none("INSERT INTO users (username,password,result_grand_total,carbon_emission_goal) VALUES ($1,$2,$3,$4)", [req.body.username, req.body.password, 0, req.body.emission])
-        const data = await db.one("SELECT * FROM users WHERE username = $1 AND password = $2 ", [req.body.username, req.body.password])
+        const hashedPass = await bcrypt.hash(req.body.password, 10)
+        await db.none("INSERT INTO users (username,password,result_grand_total,carbon_emission_goal) VALUES ($1,$2,$3,$4)", [req.body.username, hashedPass, 0, req.body.emission])
+        const data = await db.one("SELECT * FROM users WHERE username = $1 AND password = $2 ", [req.body.username, hashedPass])
         console.log(data)
         res.status(200).json(data)
     }
@@ -67,6 +68,8 @@ const setEmissionGoal =async(req,res)=> {
         res.status(500).send(err)
     }
 }
+
+
 
 module.exports = {
     login,
