+++ package.json
@@ -20,8 +20,10 @@
   },
   "homepage": "https://github.com/my-carbon-tracker/my-carbon-footprint-server#readme",
   "dependencies": {
+    "bcrypt": "^5.0.1",
     "body-parser": "^1.19.0",
     "cookie-parser": "^1.4.5",
+    "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
