+++ server.js
@@ -1,10 +1,25 @@
 require('dotenv').config();
+const cors = require('cors')
 const express = require('express');
 const userRouter = require('./routes/userRoutes')
 
 
 const app = express();
-// const bodyParser = require("body-parser");
+
+////Cors
+const whitelist = ["http://localhost:3001/", "http://localhost:3000/"]
+const corsOptions = {
+    origin: function (origin, callback) {
+        if (whitelist.indexOf(origin) !== -1) {
+            callback(null, true);
+        } else {
+            callback(
+                new Error("Not allowed by CORS; origin domain needs to be added to whitelist.")
+            );
+        }
+    },
+};
+app.use(cors())
 app.use(express.json());
 // app.use(bodyParser.json());
 app.use(express.urlencoded());
